+++ docs/js/general.js
@@ -1,3 +1,21 @@
+function errorMsg(html) {
+    Swal.fire({
+        title: Branding.errorTitle,
+        html: html,
+        icon: 'error',
+        confirmButtonText: Branding.okButtonText,
+    })
+}
+
+function infoMsg(html) {
+    Swal.fire({
+        title: Branding.infoTitle,
+        html: html,
+        icon: 'info',
+        confirmButtonText: Branding.okButtonText,
+    })
+}
+
 function boot() {
     var obj = document.getElementById("dvique");
     var canvas = document.createElement("canvas");
@@ -198,12 +216,7 @@ function create_newboard() {
             document.getElementById('constraints').style.display = 'none';
         }
     } else {
-        Swal.fire({
-            title: 'LMI says:',
-            html: 'Display size must be in the range <h2 class="warn">12-90</h2>',
-            icon: 'error',
-            confirmButtonText: 'ok 🙂',
-        })
+        errorMsg('Display size must be in the range <h2 class="warn">12-90</h2>')
     }
 }
 
@@ -248,12 +261,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (nx <= gridmax['square'] && nx > 0 && ny <= gridmax['square'] && ny > 0 && space1 + space2 < ny && space3 + space4 < nx) {
                 pu = new Puzzle_square(nx, ny, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Rows/Columns Size must be in the range <h2 class="warn">1-' + gridmax['square'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Rows/Columns Size must be in the range <h2 class="warn">1-' + gridmax['square'] + '</h2>');
             }
             break;
         case "hex":
@@ -264,12 +272,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= gridmax['hex'] && n0 > 0 && space1 < n0) {
                 pu = new Puzzle_hex(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['hex'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['hex'] + '</h2>');
             }
             break;
         case "tri":
@@ -280,12 +283,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= gridmax['tri'] && n0 > 0 && space1 < n0 / 3) {
                 pu = new Puzzle_tri(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['tri'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['tri'] + '</h2>');
             }
             break;
         case "pyramid":
@@ -296,12 +294,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= gridmax['pyramid'] && n0 > 0 && space1 < n0 / 3) {
                 pu = new Puzzle_pyramid(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['pyramid'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['pyramid'] + '</h2>');
             }
             break;
         case "iso":
@@ -309,12 +302,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= gridmax['cube'] && n0 > 0) {
                 pu = new Puzzle_iso(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['iso'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['iso'] + '</h2>');
             }
             break;
         case "sudoku":
@@ -486,12 +474,7 @@ function make_class(gridtype, loadtype = 'new') {
                     pu.draw_kakurogrid();
                 }
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Rows/Columns Size must be in the range <h2 class="warn">1-' + gridmax['kakuro'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Rows/Columns Size must be in the range <h2 class="warn">1-' + gridmax['kakuro'] + '</h2>');
             }
             break;
         case "truncated_square":
@@ -499,12 +482,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= 20 && n0 > 0) {
                 pu = new Puzzle_truncated_square(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['truncated'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['truncated'] + '</h2>');
             }
             break;
         case "tetrakis_square":
@@ -512,12 +490,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= 20 && n0 > 0) {
                 pu = new Puzzle_tetrakis_square(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['tetrakis'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['tetrakis'] + '</h2>');
             }
             break;
         case "snub_square":
@@ -525,12 +498,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= 20 && n0 > 0) {
                 pu = new Puzzle_snub_square(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['snub'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['snub'] + '</h2>');
             }
             break;
         case "cairo_pentagonal":
@@ -538,12 +506,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= 20 && n0 > 0) {
                 pu = new Puzzle_cairo_pentagonal(n0, n0, size);
             } else {
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['cairo'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['cairo'] + '</h2>');
             }
             break;
         case "rhombitrihexagonal":
@@ -551,11 +514,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= 20 && n0 > 0) {
                 pu = new Puzzle_rhombitrihexagonal(n0, n0, size);
             } else {
-                Swal.fire({
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['rhombitrihex'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['rhombitrihex'] + '</h2>');
             }
             break;
         case "deltoidal_trihexagonal":
@@ -563,11 +522,7 @@ function make_class(gridtype, loadtype = 'new') {
             if (n0 <= 20 && n0 > 0) {
                 pu = new Puzzle_deltoidal_trihexagonal(n0, n0, size);
             } else {
-                Swal.fire({
-                    html: 'Side Size must be in the range <h2 class="warn">1-' + gridmax['deltoidal'] + '</h2>',
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                })
+                errorMsg('Side Size must be in the range <h2 class="warn">1-' + gridmax['deltoidal'] + '</h2>');
             }
             break;
     }
@@ -1531,14 +1486,14 @@ function advancecontrol_onoff(loadtype = "new") {
     if (!can_use_lite() || loadtype === "off") {
         // Lite Version OFF, Display all the modes
         // Display the mode break line again
-        document.getElementById("mode_break").style.display = "inline";
-        document.getElementById("mode_txt_space").style.display = "inline";
+        document.getElementById("mode_break").classList.remove('is_hidden');
+        document.getElementById("mode_txt_space").classList.remove('is_hidden');
         advancecontrol_on();
     } else {
         // Lite Version ON, so turn off extra modes
         // Remove the mode break line again
-        document.getElementById("mode_break").style.display = "none";
-        document.getElementById("mode_txt_space").style.display = "none";
+        document.getElementById("mode_break").classList.add('is_hidden');
+        document.getElementById("mode_txt_space").classList.add('is_hidden');
         advancecontrol_off(loadtype);
     }
 }
@@ -1552,14 +1507,14 @@ function advancecontrol_off(loadtype) {
     }
 
     if (user_choices.indexOf("Surface") === -1) {
-        document.getElementById("mo_surface_lb").style.display = "none";
+        document.getElementById("mo_surface_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Line Normal") === -1 &&
         user_choices.indexOf("Line Diagonal") === -1 &&
         user_choices.indexOf("Line Free") === -1 &&
         user_choices.indexOf("Line Middle") === -1 &&
         user_choices.indexOf("Line Helper") === -1) {
-        document.getElementById("mo_line_lb").style.display = "none";
+        document.getElementById("mo_line_lb").classList.add('is_hidden');
     } else {
         // document.getElementById("st_line80_lb").style.display = "none";
         // document.getElementById("st_line12_lb").style.display = "none";
@@ -1570,7 +1525,7 @@ function advancecontrol_off(loadtype) {
         user_choices.indexOf("Edge Diagonal") === -1 &&
         user_choices.indexOf("Edge Free") === -1 &&
         user_choices.indexOf("Edge Helper") === -1) {
-        document.getElementById("mo_lineE_lb").style.display = "none";
+        document.getElementById("mo_lineE_lb").classList.add('is_hidden');
     } else {
         // document.getElementById("st_lineE80_lb").style.display = "none";
         // document.getElementById("st_lineE12_lb").style.display = "none";
@@ -1578,7 +1533,7 @@ function advancecontrol_off(loadtype) {
         // document.getElementById("st_lineE21_lb").style.display = "none";
     }
     if (user_choices.indexOf("Wall") === -1) {
-        document.getElementById("mo_wall_lb").style.display = "none";
+        document.getElementById("mo_wall_lb").classList.add('is_hidden');
     } else {
         // document.getElementById("st_wall1_lb").style.display = "none";
         // document.getElementById("st_wall12_lb").style.display = "none";
@@ -1592,32 +1547,32 @@ function advancecontrol_off(loadtype) {
         user_choices.indexOf("Number 1/4") === -1 &&
         user_choices.indexOf("Number Side") === -1 &&
         user_choices.indexOf("Candidates") === -1) {
-        document.getElementById("mo_number_lb").style.display = "none";
+        document.getElementById("mo_number_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Shape") === -1) {
-        document.getElementById("mo_symbol_lb").style.display = "none";
+        document.getElementById("mo_symbol_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Special") === -1 &&
         user_choices.indexOf("Thermo") === -1 &&
         user_choices.indexOf("Sudoku Arrow") === -1) {
-        document.getElementById("mo_special_lb").style.display = "none";
+        document.getElementById("mo_special_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Cage") === -1) {
-        document.getElementById("mo_cage_lb").style.display = "none";
+        document.getElementById("mo_cage_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Composite") === -1) {
-        document.getElementById("mo_combi_lb").style.display = "none";
+        document.getElementById("mo_combi_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Sudoku Normal") === -1 &&
         user_choices.indexOf("Sudoku Corner") === -1 &&
         user_choices.indexOf("Sudoku Centre") === -1) {
-        document.getElementById("mo_sudoku_lb").style.display = "none";
+        document.getElementById("mo_sudoku_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Box") === -1) {
-        document.getElementById("mo_board_lb").style.display = "none";
+        document.getElementById("mo_board_lb").classList.add('is_hidden');
     }
     if (user_choices.indexOf("Move") === -1) {
-        document.getElementById("mo_move_lb").style.display = "none";
+        document.getElementById("mo_move_lb").classList.add('is_hidden');
     }
 }
 
@@ -1625,7 +1580,9 @@ function advancecontrol_on() {
     if (document.getElementById('tab-dropdown-lite-btn')) {
         document.getElementById('tab-dropdown-lite-btn').innerText = "Enable Penpa Lite";
     }
-    pu.erase_buttons();
+
+    // pu.erase_buttons();
+    PenpaUI.set_visible_modes_by_gridtype(pu.gridtype);
 
     // Set the solve mode
     if (pu.mmode === "solve") {
@@ -1863,12 +1820,7 @@ function saveimage_download() {
             }
         }
     } else {
-        Swal.fire({
-            title: 'LMI says:',
-            html: 'The characters <h2 class="warn">\\ / : * ? \" < > |</h2> cannot be used in filename',
-            icon: 'error',
-            confirmButtonText: 'ok 🙂',
-        })
+        errorMsg('The characters <h2 class="warn">\\ / : * ? \" < > |</h2> cannot be used in filename');
     }
 }
 
@@ -2317,11 +2269,7 @@ function validate_entries() {
 }
 
 function savetext_copy() {
-    Swal.fire({
-        title: 'LMI says:',
-        html: '<h2 class="info">URL is copied to clipboard</h2>',
-        icon: 'info'
-    })
+    infoMsg('<h2 class="info">URL is copied to clipboard</h2>');
     var textarea = document.getElementById("savetextarea");
     textarea.select();
     var range = document.createRange();
@@ -2374,12 +2322,7 @@ function saveblob_download(blob, defaultFilename) {
             });
         }
     } else {
-        Swal.fire({
-            title: 'LMI says:',
-            html: 'The characters <h2 class="warn">\\ / : * ? \" < > |</h2> cannot be used in filename',
-            icon: 'error',
-            confirmButtonText: 'ok 🙂',
-        });
+        errorMsg('The characters <h2 class="warn">\\ / : * ? \" < > |</h2> cannot be used in filename');
     }
 }
 
@@ -2528,10 +2471,7 @@ function load_feedback() {
 }
 
 function show_shortcuts() {
-    Swal.fire({
-        title: 'Shortcuts',
-        html: '<table style="width:100%" class="shortcuts"><tr><th style="color:Green;" colspan="2">General</th></tr><tr><th style="color:red;">F2</th><th>Problem mode</th></tr><tr><th style="color:red;">F3</th><th>Solution mode</th></tr><tr><th style="color:red;">F4</th><th>Hide/Show Timer</th></tr><tr><th style="color:red;">Ctrl + D</th><th>Clone/Duplicate</th></tr><tr><th style="color:Green;" colspan="2">Sudoku Mode</th></tr><tr><th style="color:red;">Z</th><th>Normal Submode</th></tr><tr><th style="color:red;">X</th><th>Corner Submode</th></tr><tr><th style="color:red;">C</th><th>Centre Submode</th></tr><tr><th style="color:red;">V</th><th>Shading (Surface Mode)</th></tr><tr><th style="color:red;">SHIFT</th><th>For Temporary Corner Submode</th></tr><tr><th style="color:red;">SHIFT + DEL</th><th>Deletes only corner pencil marks from the selected cells</th></tr><tr><th style="color:red;">CTRL</th><th>For Temporary Centre Submode</th></tr><tr><th style="color:red;">CTRL + DEL</th><th>Deletes only centre pencil marks from the selected cells</th></tr><tr><th style="color:red;">Border: ON</th><th>will allow you to write digits on the edges</th></tr><tr><th style="color:Green;" colspan="2">Surface Mode</th></tr><tr><th colspan="2">Use number keys to switch between styles</th></tr> </table>',
-    })
+    document.getElementById("modal-keys").style.display = 'block';
 }
 
 function load(urlParam, type = 'url', origurl = null) {
@@ -2907,7 +2847,7 @@ function load2(paramArray, type, origurl) {
 
     // Set some genre specific settings
     if ($('#genre_tags_opt').select2("val").includes("alphabet")) {
-        UserSettings.disable_shortcuts = 2;
+        UserSettings.shortcuts_enabled = false;
     }
 
     if (paramArray.m === "edit") { //edit_mode
@@ -3759,7 +3699,7 @@ function set_solvemode(type = "url") {
     document.getElementById("pu_a_label").style.display = "none";
     document.getElementById("newboard").style.display = "none";
     document.getElementById("rotation").style.display = "none";
-    document.getElementById("mo_board_lb").style.display = "none";
+    document.getElementById("mo_board_lb").classList.add('is_hidden');
     document.getElementById("sub_number2_lb").style.display = "none";
     document.getElementById("sub_number4_lb").style.display = "none";
     document.getElementById("sub_number11_lb").style.display = "none";
@@ -3821,7 +3761,7 @@ function set_contestmode() {
     document.getElementById("tb_reset").style.display = "none";
     document.getElementById("tb_delete").style.display = "none";
     document.getElementById("tb_delete_top").style.display = "inline-block";
-    document.getElementById("mo_move_lb").style.display = "none";
+    document.getElementById("mo_move_lb").classList.add('is_hidden');
     document.getElementById("puzzlesourcelink").style.display = "none";
     document.getElementById("saveinfotitle_lb").style.display = "none";
     document.getElementById("saveinfotitle").style.display = "none";
@@ -3843,19 +3783,9 @@ function display_answercheck() {
     // Validate at least one answer check option is selected
     var answer_check_opt = pu.get_answercheck_settings();
     if (answer_check_opt.answercheck_opt.length === 0) {
-        Swal.fire({
-            title: 'LMI says:',
-            html: 'No specific option selected by Author. Answer check looks for all the elements with appropriate accepted colors. Check <a href="https://github.com/swaroopg92/penpa-edit/blob/master/images/multisolution.PNG" target="_blank">this</a> for reference.',
-            icon: 'info',
-            confirmButtonText: 'ok 🙂',
-        })
+        infoMsg('No specific option selected by Author. Answer check looks for all the elements with appropriate accepted colors. Check <a href="https://github.com/swaroopg92/penpa-edit/blob/master/images/multisolution.PNG" target="_blank">this</a> for reference.');
     } else {
-        Swal.fire({
-            title: 'LMI says:',
-            html: answer_check_opt.message,
-            icon: 'info',
-            confirmButtonText: 'ok 🙂',
-        })
+        infoMsg(answer_check_opt.message);
     }
 }
 
@@ -3892,12 +3822,7 @@ function decode_puzzlink(url) {
     rows = parseInt(urldata[2]);
 
     if ((cols > 65) || (rows > 65)) {
-        Swal.fire({
-            title: 'LMI says:',
-            html: 'Penpa+ does not support grid size greater than 65 rows or columns',
-            icon: 'error',
-            confirmButtonText: 'ok 🙂',
-        })
+        errorMsg('Penpa+ does not support grid size greater than 65 rows or columns');
         return;
     }
 
@@ -3994,12 +3919,7 @@ function decode_puzzlink(url) {
             } else {
                 pu = new Puzzle_square(10, 10, size);
                 setupProblem(pu, "surface");
-                Swal.fire({
-                    title: 'LMI says:',
-                    html: `Sorry, sudoku grids of size: ${cols}x${rows} are not supported`,
-                    icon: 'error',
-                    confirmButtonText: 'ok 🙂',
-                });
+                errorMsg(`Sorry, sudoku grids of size: ${cols}x${rows} are not supported`);
                 break;
             }
             setupProblem(pu, "sudoku");
@@ -5610,19 +5530,14 @@ function decode_puzzlink(url) {
             pu.user_tags = ["rassi silai"]; // Genre Tags
             break;
         default:
-            Swal.fire({
-                title: 'LMI says:',
-                html: 'It currently does not support puzzle type: ' + type,
-                icon: 'error',
-                confirmButtonText: 'ok 🙂',
-            })
+            errorMsg('It currently does not support puzzle type: ' + type);
             break;
     }
 
     // Set PenpaLite
     // document.getElementById('advance_button').value = "1";
-    document.getElementById("mode_break").style.display = "none";
-    document.getElementById("mode_txt_space").style.display = "none";
+    document.getElementById("mode_break").classList.add('is_hidden');
+    document.getElementById("mode_txt_space").classList.add('is_hidden');
     advancecontrol_off("url");
 
     var tabSelect = document.querySelector('ul.multi');