+++ docs/js/settings.js
@@ -218,15 +218,14 @@ const UserSettings = {
         return this._reload_button;
     },
 
-    _disable_shortcuts: 1,
-    set disable_shortcuts(newValue) {
-        const valueInt = newValue ? parseInt(newValue, 10) : 1;
-        this._disable_shortcuts = valueInt;
+    _shortcuts_enabled: 1,
+    set shortcuts_enabled(newValue) {
+        this._shortcuts_enabled = newValue;
 
-        document.getElementById("disable_shortcuts_opt").value = valueInt;
+        document.getElementById("enable_shortcuts_opt").value = newValue ? 1 : 2;
     },
-    get disable_shortcuts() {
-        return this._disable_shortcuts;
+    get shortcuts_enabled() {
+        return this._shortcuts_enabled;
     },
 
     _gridtype: "square",
@@ -284,21 +283,11 @@ const UserSettings = {
 
         if (valueInt > 90) {
             valueInt = 90;
-            Swal.fire({
-                title: 'LMI says:',
-                html: 'Display Size must be in the range <h2 class="warn">12-90</h2> It is set to max value.',
-                icon: 'info',
-                confirmButtonText: 'ok 🙂',
-            })
+            infoMsg('Display Size must be in the range <h2 class="warn">12-90</h2> It is set to max value.');
         }
         if (valueInt < 12) {
             valueInt = 12;
-            Swal.fire({
-                title: 'LMI says:',
-                html: 'Display Size must be in the range <h2 class="warn">12-90</h2> It is set to min value.',
-                icon: 'info',
-                confirmButtonText: 'ok 🙂',
-            })
+            infoMsg('Display Size must be in the range <h2 class="warn">12-90</h2> It is set to min value.');
         }
 
         this._displaysize = valueInt;
@@ -423,12 +412,7 @@ const UserSettings = {
         deleteCookie('tab_settings');
         // deleteCookie("different_solution_tab");
 
-        Swal.fire({
-            title: 'Cookies cleared!',
-            html: 'You must reload the page for the default settings to take effect.',
-            icon: 'info',
-            confirmButtonText: 'ok 🙂',
-        });
+        infoMsg('You must reload the page for the default settings to take effect.');
     },
 
     _settingsLoaded: false,