+++ CHANGELOG.md
@@ -1,7 +1,4 @@
 ## History
-* unreleased
-	* Improved sudoku string IO to add whitespace suppression option and import non-square text blocks.
-	* Separated branding for easy access in future to other applications.
 * 2023/07/28 ver 3.0.8
 	* Updated jquery, sweetalert to latest.
 	* Fixed the bug of correctly deleteing cagearrays.
@@ -11,6 +8,8 @@
 	* Improved shortcuts display under Help, adding ? to access Help.
 	* Fixed ? in tapa gmpuzzles output text, minor display improvements.
 	* Code cleanup
+	* Improved sudoku string IO to add whitespace suppression option and import non-square text blocks.
+	* Separated branding for easy access in future to other applications.
 * 2023/05/03 ver 3.0.7
 	* Diagonal connection shape for killer cages with rotation fix.
 	* Tweaked display for mobile devices.